Restructured API routes
