+++ src/pages/api/download-reports/ror/funders.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next"
 import { gql } from '@apollo/client';
-import apolloClient from '../../../utils/apolloClient'
+import apolloClient from '../../../../utils/apolloClient'
 import { stringify } from 'csv-stringify/sync'
 
 const QUERY = gql`
@@ -23,7 +23,7 @@ const QUERY = gql`
       crossrefFunderId: $crossrefFunderId
     ) {
       works(
-        first: 200
+        first: 0
         after: $cursor
         query: $filterQuery
         published: $published
@@ -32,24 +32,15 @@ const QUERY = gql`
         language: $language
         license: $license
         registrationAgency: $registrationAgency
+        facetCount: 200
       ) {
-        nodes {
-          ...WorkFragment
+        funders {
+          id
+          title
+          count
         }
       }
     }
-  }
-  fragment WorkFragment on Work {
-    titles {
-      title
-    }
-    descriptions {
-      description
-      descriptionType
-    }
-    doi
-    formattedCitation(style: "apa", locale: "en-US", format: text)
-    publicationYear
   }
 `
 
@@ -64,27 +55,20 @@ export default async function downloadReportsHandler(
     query: QUERY,
     variables: variables
   })
-
-  const sortedData = [...data.organization.works.nodes].sort((a, b) => b.publicationYear - a.publicationYear)
-
-  const csv = stringify(sortedData, {
-    header: true,
-    columns: [
-      { key: 'titles[0].title', header: 'Title' },
-      { key: 'publicationYear', header: 'Publication Year' },
-      { key: 'doi', header: 'DOI' },
-      { key: 'descriptions[0].description', header: 'Description' },
-      { key: 'formattedCitation', header: 'Formatted Citation' }
-    ]
-  })
+  
 	
+  const csv = stringify(data.organization.works.funders, {
+    header: true,
+    columns: [ { key: 'id', header: 'Funder ID' }, { key: 'title', header: 'Title' }, { key: 'count', header: 'Work Count' } ]
+  })
+
+
   try {
     res.status(200)
     res.setHeader('Content-Type', 'text/csv')
-    res.setHeader('Content-Disposition', `attachment; filename="related-works_${variables.id}.csv"`)
+    res.setHeader('Content-Disposition', `attachment; filename="funders_${variables.id}.csv"`)
     res.send(csv)
   } catch (error) {
     res.status(400).json({ error })
   }
-
 }
