+++ src/pages/api/download-reports/ror/related-works.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next"
 import { gql } from '@apollo/client';
-import apolloClient from '../../../utils/apolloClient'
+import apolloClient from '../../../../utils/apolloClient'
 import { stringify } from 'csv-stringify/sync'
 
 const QUERY = gql`
@@ -23,7 +23,7 @@ const QUERY = gql`
       crossrefFunderId: $crossrefFunderId
     ) {
       works(
-        first: 0
+        first: 200
         after: $cursor
         query: $filterQuery
         published: $published
@@ -32,15 +32,29 @@ const QUERY = gql`
         language: $language
         license: $license
         registrationAgency: $registrationAgency
-        facetCount: 200
       ) {
-        funders {
-          id
-          title
-          count
+        nodes {
+          ...WorkFragment
         }
       }
     }
+  }
+  
+  fragment WorkFragment on Work {
+    titles {
+      title
+    }
+    descriptions {
+      description
+      descriptionType
+    }
+    types {
+      resourceTypeGeneral
+      resourceType
+    }
+    doi
+    formattedCitation(style: "apa", locale: "en-US", format: text)
+    publicationYear
   }
 `
 
@@ -55,20 +69,30 @@ export default async function downloadReportsHandler(
     query: QUERY,
     variables: variables
   })
-  
-	
-  const csv = stringify(data.organization.works.funders, {
+
+  const sortedData = [...data.organization.works.nodes].sort((a, b) => b.publicationYear - a.publicationYear)
+
+  const csv = stringify(sortedData, {
     header: true,
-    columns: [ { key: 'id', header: 'Funder ID' }, { key: 'title', header: 'Title' }, { key: 'count', header: 'Work Count' } ]
+    columns: [
+      { key: 'titles[0].title', header: 'Title' },
+      { key: 'publicationYear', header: 'Publication Year' },
+      { key: 'doi', header: 'DOI' },
+      { key: 'descriptions[0].description', header: 'Description' },
+      { key: 'formattedCitation', header: 'Formatted Citation' },
+      { key: 'types.resourceTypeGeneral', header: 'Resource Type (General)' },
+      { key: 'types.resourceType', header: 'Resource Type' },
+      // { key: '???', header: 'Connection Type(s)' }
+    ]
   })
-
-
+	
   try {
     res.status(200)
     res.setHeader('Content-Type', 'text/csv')
-    res.setHeader('Content-Disposition', `attachment; filename="funders_${variables.id}.csv"`)
+    res.setHeader('Content-Disposition', `attachment; filename="related-works_${variables.id}.csv"`)
     res.send(csv)
   } catch (error) {
     res.status(400).json({ error })
   }
+
 }
